pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "neon_gridiron"
version = "0.2.0"
description = "Professional 7v7 Futuristic Football simulation with MARL"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Antigravity", email = "ai@google.com"}
]
dependencies = [
    "numpy>=1.26.0",
    "pymunk>=6.6.0",
    "pygame>=2.5.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "torch>=2.2.0",
    "gymnasium>=0.29.0",
    "stable-baselines3>=2.2.1",
    "shimmy>=1.3.0",
    "pydantic>=2.6.0",
    "pyyaml>=6.0.1",
    "msgpack>=1.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
    "pre-commit>=3.6.0",
    "black>=24.2.0",
]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "B", "I", "U"]
ignore = ["E501"] # Line too long handled by formatter

[tool.mypy]
python_version = "3.12"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
